+++ exercise_1.py
@@ -1,7 +1,0 @@
-milk = 2.50
-bread= 1.25
-eggs = 2.00
-
-
-total_cost = milk + bread + eggs
-print(f"Your total cost is ${total_cost}.")