+++ CHANGELOG.md
@@ -1,4 +1,17 @@
 # Changelog
+day-5-functions-and-modules
+
+## [0.6.0]
+## -2024-29-24
+### Added
+- agl_core.py: add functions to add, edit, list, and remove items from a grocery list
+- Agl_launch.py:  add function to interact with user to add, edit, list, and remove items from a grocery list
+- added practice exercise to my_script.py
+
+## [0.5.0] -2024-11-24
+### Added 
+-Day 4 Assignment
+=======
 
 ## [0.5.0] -2024-11-24
 ### Added 
@@ -21,6 +34,24 @@ Loop Control, and Error Messages
 ### Initial Version
 - Added core functionalities covered in Day 3, including lists, tuples, dictionaries, and sets.
 - Implemented changelog and version control instructions.
+
+
+
+
+day-5-functions-and-modules
+## [0.3.0] -2024-11-24
+### Added
+-exercises in for and while loops
+
+## [0.2.0] -2024-11-23
+### Added
+- exercises for tuples, dictionaries, sets, and slicing lists
+
+## [0.1.0] - 2024-11-23
+### Initial Version
+- Added core functionalities covered in Day 3, including lists, tuples, dictionaries, and sets.
+- Implemented changelog and version control instructions.
+